mudanças do renanzinho/ common uses
